Attempt to enable metrics from upstream k8s helm
No experience with helm charts or prometheus, in an attempt to learn about them I saw
that the k8s-at-home sonarr chart had a feature for metrics through
exportarr. My assumption is that since truecharts uses k8s-at-home
as a base for the chart that I'd be able to copy and paste the
portions of templates and values and questions to have it hopefully all
work out.